16.10.2003 Large patch from Paul included, updated German translation
git-svn-id: svn://scribus.net/trunk/Scribus@80 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/align.h
@@ -46,7 +46,7 @@ public slots:
     virtual void slotApplyDiag();
 
 signals:
-		void ApplyDist(bool, bool, bool, bool, float, float, int, int);
+		void ApplyDist(bool, bool, bool, bool, double, double, int, int);
 
 protected:
     QVBoxLayout* AlignLayout;
