+++ scribus/annota.h
@@ -56,7 +56,7 @@ public:
 		int MaxSeite;
 
 public slots:
-		void SetCo(float x, float y);
+		void SetCo(double x, double y);
 		void SetPg(int v);
 		void SetCross();
 		void SetVals();
