+++ scribus/hruler.cpp
@@ -54,8 +54,8 @@ void Hruler::mouseMoveEvent(QMouseEvent *m)
 void Hruler::paintEvent(QPaintEvent *)
 {
 	int pc, xx;
-	float of, xl, iter, iter2;
-	float sc = doku->Scale;
+	double of, xl, iter, iter2;
+	double sc = doku->Scale;
 	switch (doku->Einheit)
 		{
 		case 0:
