+++ scribus/mdup.cpp
@@ -8,9 +8,9 @@
 #include <qvariant.h>
 #include <qtooltip.h>
 extern QPixmap loadIcon(QString nam);
-extern float UmReFaktor;
+extern double UmReFaktor;
 
-Mdup::Mdup( QWidget* parent, float Dx, float Dy, int Ein )
+Mdup::Mdup( QWidget* parent, double Dx, double Dy, int Ein )
     : QDialog( parent, "m", true, 0 )
 {
     setCaption( tr( "Multiple Duplicate" ) );
